requirements.txt
Django
line-bot-sdk
configparser
gunicorn